fix: Fix syntax issues

+++ star-dash/star-dash/Rendering/OverlayInfo.swift
@@ -1,3 +1,3 @@
 struct OverlayInfo {
-    let score: Float
+    let score: Int
 }
